refactor: asset크기 최적화 및 hooks를 통해 로직 분리, 검색결과 캐싱
- gif 대신 video태그와 mp4이용
- png대신 압축이 적용된 jpg이용
- 검색어를 key 결과를 value로  메모리에 캐싱
