+++ README.md
@@ -1,49 +1,10 @@
 # 🚀 프론트엔드 성능 베이스캠프
 ## 성능 오답노트 : memegle 프로젝트 성능 개선하기
-
-## 📕 미션 소개
-
-프론트엔드 성능 베이스캠프에 오신 여러분 환영합니다! 🤗
-
-이번 미션에서는 예제로 구성한 짤 검색 사이트, 'Memegle' 프로젝트의 성능을 개선해볼거에요.     
-    
 <img width=400 src="https://user-images.githubusercontent.com/81607552/129674696-2fe7251b-90fe-4dec-8bc5-5d47bcc9159c.png"> <img width=370 src="https://user-images.githubusercontent.com/81607552/129674723-03d93732-1aba-42ca-a7cf-d2abe1005847.png">
-    
-Memegle 프로젝트는 곳곳에 성능을 저하시키는 요소들로 가득한데요.     
-여기저기 구멍난 곳들을 고쳐서, 기본적인 수준으로 쓰는 데에 불편함이 없는 버전 1.0.0을 만들어주세요. 
-
-(엉망진창 성능의) 배포된 샘플 프로젝트는 [여기](http://frontend-performance-basecamp.s3-website.ap-northeast-2.amazonaws.com)에서 확인할 수 있습니다.
-
-## ✅ 미션을 시작하기 전에...
-
-몇 가지 계정 생성이 필요합니다.
-
-- gif 검색을 위해 GIPHY API 를 사용하고 있습니다.    
-  GIPHY API 키 발급을 위해 [GIPHY](https://developers.giphy.com/)에서 가입 후 앱 생성 > API 키 발급이 필요합니다.
-    - 발급한 키는 루트 경로에 `.env`파일을 만들어 GIPHY_API_KEY 값으로 설정해주세요
-
-    ```xml
-    // .env
-    GIPHY_API_KEY={발급받은 API KEY}
-    ```
-
-- 우테코 AWS 개인 계정에서 진행합니다. 
-
-
-## 📅 **진행 방식**
-
-- 이번 미션은 개인 미션입니다. 페어는 리뷰 진행을 위해 매칭합니다.
-- 예제 프로젝트를 fork 받아 각자 개선 작업을 진행하신 뒤, PR을 보내주세요.
-- PR 제출시, PR 템플릿에 있는 항목들을 확인하고 채워주세요.
-- 페어가 제출한 PR을 상호리뷰해주세요.
 
 ## **🎯 요구사항**
 
 ### 개선 목표
-
-아래 작업 목록을 다 완료했을 경우 달성할 수 있을 것으로 예상되는 로딩 성능 관련 수치입니다.     
-측정 환경에 따라 조금씩 다를 수도 있으니, 참고 기준으로만 삼아주세요     
-이 수치를 완벽히 맞추기 위한 미션이 아니라, 개선 방법들을 한번씩 직접 경험해보기 위한 미션이랍니다. 🙂
 
 - Lighthouse `95점` 이상
 - Home 페이지에서 불러오는 스크립트 리소스 크기 `< 60kb`
