+++ webpack.config.js
@@ -2,7 +2,7 @@ const path = require("path");
 const HtmlWebpackPlugin = require("html-webpack-plugin");
 const Dotenv = require("dotenv-webpack");
 
-const isProduction = process.env.NODE_ENV;
+const isProduction = process.env.NODE_ENV === "production";
 
 module.exports = {
   entry: "./src/index.js",
@@ -15,7 +15,6 @@ module.exports = {
   ...(!isProduction && {
     devServer: {
       hot: true,
-      open: true,
       historyApiFallback: true,
     },
   }),
@@ -40,7 +39,7 @@ module.exports = {
         use: ["style-loader", "css-loader"],
       },
       {
-        test: /\.(eot|svg|ttf|woff|woff2|png|jpg|gif)$/i,
+        test: /\.(mp4|eot|svg|ttf|woff|woff2|png|jpg|gif)$/i,
         type: "asset/resource",
       },
     ],
